pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "setuptools_scm[toml]>=7.1.0"]
build-backend = "setuptools.build_meta"

[project]
name = "GlueMP4"
authors = [
  { name="Toon Verstraelen", email="toon.verstraelen@ugent.be" },
]
description = "Wrapper for ffmpeg to trim, crop and concatenate MP4 video fragments."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Environment :: Console",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: POSIX",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "pyyaml",
]
dynamic = ["version"]

[project.scripts]
gluemp4 = "gluemp4.__main__:main"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
select = ["E", "F", "UP", "B", "I", "PGH", "PL", "RUF", "C"]
line-length = 100
target-version = "py311"
ignore = ["PLR2004", "PLR0913", "PLR0912", "PLW2901", "PLR0915", "C901", "PLR0911", "RUF001"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.setuptools_scm]
write_to = "src/gluemp4/_version.py"
version_scheme = "post-release"
local_scheme = "no-local-version"
